add staging feature class and move module registration into it
